requirements.txt
beautifulsoup4==4.8.2
Django==1.11.11
Pillow==6.2.1
PyMySQL==0.9.3
pytz==2019.3
soupsieve==1.9.5
